grapeguard-ml/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2

